make alt+backspace work in vscode by default
On MacOS, vscode generates control-alt-h in response to alt+backspace.
This is handled in bash as cut-smallword-left and in zsh as cut-word-left
(zsh doesn't seem to have a concept of small/big words). Fish treats this
as cut small word left (fish bigwords include anything up to /, and its
small words stop on any punctuation).