Reduce use of partial functions in parser code
By replacing spans[idx] with spans.get(idx) calls, we can guarantee
that we won't get any Index Out of Bounds panics.

Some of the ones I replaced are very obviously checked, but that means
that I can replace a check by a pattern match. And the goal is to remove
all unchecked indexing in the parser.