Fix exponential parser time on sequence of [[[[
The old code first tried parsing the head of the table as a
list and then after that it checked if it got more arguments.
If it didn't, it throws away the previous result and tries to
parse the whole thing as a list, which means we call
parse_list_expression twice for each call to
parse_table_expression, resulting in the exponential growth

The fix is to simply check that we have all the arguments
we need before parsing the head of the table, so we know
that we will either call parse_list_expression only on
sub-expressions or on the whole thing, never both.

Fixes #10438