Prevent qubic time on nested parentheses
When parse_range get an item like ((((1..2))))
it would try to parse "((((1" with a long chain of
recursive parsers, namely:
- parse_block
- parse_pipeline
- parse_builtin_commands
- parse_expression
- parse_math_expression
- parse_value
- parse_paren_expr
- parse_full_cell_path

where parse_paren_expr calls parse_range in turn. Because at any time in
the chain `parse_paren_expr` can call `parse_range`, which will then
continue the chain, we get quadratic number of function calls,
each linear on the size of the input

By checking with the lexer that the parens are matched, we
prevent the long chain from being called on unmatched braces.
Now, this is still more quadratic than it needs to be, to fix that,
we should process parens only once, instead of on each recursive call