expand record! macro
This expands the syntax of the `record!` macro. In addition to the `key
=> value` syntax, one can now also put a simple identifier for a column.
This creates a column with the same name as that of the identifier, and
a value of the identifier's value.

This makes it easier to make a record when you already have a variable
with the value you want and want to give you column the same name as
that of the variable. e.g.

```rust
let foo = Value::int(1, Span::unknown());
let bar = Value::int(2, Span::unknown());
let rec = record!(foo, bar);
```

This will create a record like:

```nu
{ foo: 1, bar: 2 }
```