Spell out our platform support policy
This reflects how we currently distributed our efforts to supporting the
different platforms and clarifies which things get run by the Nushell
team and which rely on the help of third-parties or individual
contributors.

Feel free to propose improvements, as long as they can be backed up by
implemented practice as a result.