Remove registry clean_string hack
This was a workaround for a [bug][gentoo90/winreg-rs#52] in winreg which
has since [been fixed][edf9eef] and released in [winreg v0.12.0].

winreg now properly displays strings in RegKey's Display impl instead of
outputting their debug representation. We remove our `clean_string` such
that registry entries which happen to start/end with " or contain \\
won't get mangled. This is very important for entries in UNC path format
as those begin with a double backslash.

[gentoo90/winreg-rs#52]: <https://github.com/gentoo90/winreg-rs/issues/52>
[edf9eef]: <https://github.com/gentoo90/winreg-rs/commit/edf9eef38fd6bda9f8acb6dfea9b846738c29ff0>
[winreg v0.12.0]: <https://github.com/gentoo90/winreg-rs/releases/tag/v0.12.0>