Add `Record.insert()` that respects existing keys
The existing `Record.push()` doesn't take into account if a key already
exists. The second insertion could become inaccessible and the data
meaningless. For most purposes `.insert()` is the better choice.

It returns `Some(Value)` with the old value if found.

As we limit ourselves to `String` keys, we can use a `Cow<str>` to only
clone the key if we need to.