Use `Record::into_values` for `values` command
Note this currently makes the strong assumption that `LazyRecord::collect` which
is a trait method with default implementation always returns a
`Value::Record`. As implementers can override that default
implementation, this may not hold.

Todo: For more strict type safety change `LazyRecord::collect` to return just a `Record` (span can be
obtained separately)