Rewrite `transpose` record manipulation to new API
As part of that, we can make the overwriting with `keep-all` and
`keep-last` inplace so the items don't need to be removed and repushed
to the record. Doing so eliminates extra clones.

This changes the order of apperance in the transposed record. Now the
order is determined by the first appearance and not by the last
appearance in the ingoing columns.
This mirrors the behavior when not passed `keep-all` or `keep-last`.

Sadly the `transpose` command is so far undertested for more complex
operations.

Generally elide a bunch of unnecessary clones.