Add naive `Record.retain()` for inplace removal
This is suboptimal to write with two independent `Vec`s as `Vec::retain`
takes care of minimizing the shifting burden internally.
For the best behavior we can get with that use `retain` for the larger
`vals` and `remove` for `cols`.

I was considering allocating a `Vec<usize>` storing indices to do a
second `retain` pass on `cols` but this makes some assumptions about the
regular use of `Record::retain` I didn't want to make:
- Should we store keep or remove indices?
- The allocation cost would be high if the first question was answered
unfavorably.
- The general overhead would be high if the regular use would just hit a
few sparse removals.

Don't have a guaranteed use in mind yet. (probably in most cases
`remove` will be used or non inplace solutions are still prevalent)