Improve case insensitivity consistency
unicase is already a dependency via mime_guess in nu-command. This
replaces as much ad-hoc lowercase/uppercase comparison as possible with
case folding provided by unicase.

In actuality a lot of code still does to_lowercase, because unicase only
provides immediate comparison, and doesn't expose a to_folded_case yet.
And since we do a lot of contains/starts_with/ends_with, it's not enough
to just have eq_ignore_case. But if we get access, this makes us ready
to use it with a change in one place.

Plus, it's clearer what the purpose is at the call site to call
to_folded_case instead of to_lowercase if it's just for the purpose of
case insensitive comparison, even if it just does to_lowercase still.