Provide configuration to plugin calls
When nushell calls a plugin `nu_plugin_NAME` it now sends a configuration
`Value` from the nushell config under `$env.config.plugins.NAME`.  This
allows plugin authors to read configuration provided by plugin users.

The `NAME` must match the registered plugin filename following `nu_plugin_`.
If you register `target/debug/nu_plugin_config` the `NAME` will be
`config` and the nushell config will loook like:

    $env.config = {
      # ...
      plugins: {
        config: [
          some
          values
        ]
      }
    }

This is a breaking change for the plugin API as the Plugin::run()
function now accepts a new configuration argument which is an
`&Option<Value>`.  If no configuration was supplied the value is `None`.

Plugins compiled after this change should work with older nushell, and
will behave as if the configuration was not set.