Add Argument::span() and Call::arguments_span()
These make it easy to make a Span that covers an entire argument and the span
of all arguments in a Call.

Call::arguments_span() is useful for errors where a command may accept
arguments or the pipeline, but not both.

Argument::span() is useful for errors where an arguments is incompatible
with one or more other arguments.